services/app/app_requirements.txt
langchain==0.2.16
langchain_experimental==0.0.65
langchain-openai==0.1.25
pydantic-settings==2.5.2
langchain_ollama==0.1.3
ollama==0.3.2
pandas
pyarrow==17.0.0
sentence_transformers==3.1.1
elasticsearch
pytest 
pytest-flask
Flask
mysql-connector